+++ Tests/TuistKitTests/ProjectEditor/Mocks/MockProjectEditorMapper.swift
@@ -7,13 +7,15 @@ import TuistCore
 
 final class MockProjectEditorMapper: ProjectEditorMapping {
     var mapStub: (Project, Graph)?
-    var mapArgs: [(sourceRootPath: AbsolutePath, manifests: [AbsolutePath], helpers: [AbsolutePath], projectDescriptionPath: AbsolutePath)] = []
+    var mapArgs: [(tuistPath: AbsolutePath, sourceRootPath: AbsolutePath, manifests: [AbsolutePath], helpers: [AbsolutePath], projectDescriptionPath: AbsolutePath)] = []
 
-    func map(sourceRootPath: AbsolutePath,
+    func map(tuistPath: AbsolutePath,
+             sourceRootPath: AbsolutePath,
              manifests: [AbsolutePath],
              helpers: [AbsolutePath],
              projectDescriptionPath: AbsolutePath) -> (Project, Graph) {
-        mapArgs.append((sourceRootPath: sourceRootPath,
+        mapArgs.append((tuistPath: tuistPath,
+                        sourceRootPath: sourceRootPath,
                         manifests: manifests,
                         helpers: helpers,
                         projectDescriptionPath: projectDescriptionPath))