+++ tools/tuistbench/Sources/TuistBenchmark/Utils/Array+Double.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension Array where Element == Double {
+    func average() -> Double {
+        reduce(0, +) / Double(count)
+    }
+}